+++ pkg/integrations/install/install.go
@@ -27,5 +27,6 @@ import (
 	// v2 integrations
 	//
 
-	_ "github.com/grafana/agent/pkg/integrations/v2/agent" // register agent
+	_ "github.com/grafana/agent/pkg/integrations/v2/agent"             // register agent
+	_ "github.com/grafana/agent/pkg/integrations/v2/app_o11y_receiver" // register app_o11y_receiver
 )