+++ pkg/integrations/v2/app_o11y_receiver/exporters/exporter.go
@@ -1,0 +1,11 @@
+package exporters
+
+import (
+	"github.com/grafana/agent/pkg/integrations/v2/app_o11y_receiver/models"
+)
+
+// AppO11yReceiverExporter is the base interface
+type AppO11yReceiverExporter interface {
+	Name() string
+	Export(payload models.Payload) error
+}