+++ pkg/integrations/v2/app_o11y_receiver/handler/handler.go
@@ -1,0 +1,100 @@
+package handler
+
+import (
+	"sync"
+
+	"encoding/json"
+	"net/http"
+
+	"github.com/go-kit/log"
+	"github.com/go-kit/log/level"
+	"github.com/grafana/agent/pkg/integrations/v2/app_o11y_receiver/config"
+	"github.com/grafana/agent/pkg/integrations/v2/app_o11y_receiver/exporters"
+	"github.com/grafana/agent/pkg/integrations/v2/app_o11y_receiver/models"
+	"github.com/grafana/agent/pkg/integrations/v2/app_o11y_receiver/tools/ratelimiting"
+	"github.com/rs/cors"
+)
+
+// AppO11yHandler struct controls the data ingestion http handler of the receiver
+type AppO11yHandler struct {
+	exporters   []exporters.AppO11yReceiverExporter
+	config      config.AppO11yReceiverConfig
+	rateLimiter *ratelimiting.RateLimiter
+}
+
+// NewAppO11yHandler creates a new AppReceiver instance based on the given configuration
+func NewAppO11yHandler(conf config.AppO11yReceiverConfig, exporters []exporters.AppO11yReceiverExporter) AppO11yHandler {
+	var rateLimiter *ratelimiting.RateLimiter
+	if conf.RateLimiting.Enabled {
+		var rps float64
+		if conf.RateLimiting.RPS > 0 {
+			rps = conf.RateLimiting.RPS
+		}
+
+		var b int
+		if conf.RateLimiting.Burstiness > 0 {
+			b = conf.RateLimiting.Burstiness
+		}
+		rateLimiter = ratelimiting.NewRateLimiter(rps, b)
+	}
+
+	return AppO11yHandler{
+		exporters:   exporters,
+		config:      conf,
+		rateLimiter: rateLimiter,
+	}
+}
+
+// HTTPHandler is the http.Handler for the receiver. It will do the following
+// 0. Enable CORS for the configured hosts
+// 1. Check if the request should be rate limited
+// 2. Verify that the payload size is within limits
+// 3. Start two go routines for exporters processing and exporting data respectively
+// 4. Respond with 202 once all the work is done
+func (ar *AppO11yHandler) HTTPHandler(logger log.Logger) http.Handler {
+	var handler http.Handler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		// Check rate limiting state
+		if ar.config.RateLimiting.Enabled && ar.rateLimiter.IsRateLimited() {
+			http.Error(w, http.StatusText(http.StatusTooManyRequests), http.StatusTooManyRequests)
+			return
+		}
+
+		// Verify content length. We trust net/http to give us the correct number
+		if ar.config.MaxAllowedPayloadSize > 0 && r.ContentLength > ar.config.MaxAllowedPayloadSize {
+			http.Error(w, http.StatusText(http.StatusRequestEntityTooLarge), http.StatusRequestEntityTooLarge)
+			return
+		}
+
+		var p models.Payload
+		err := json.NewDecoder(r.Body).Decode(&p)
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusBadRequest)
+			return
+		}
+
+		var wg sync.WaitGroup
+
+		for _, exporter := range ar.exporters {
+			wg.Add(1)
+			go func(exp exporters.AppO11yReceiverExporter) {
+				defer wg.Done()
+				if err := exp.Export(p); err != nil {
+					level.Error(logger).Log("msg", "exporter error", "exporter", exp.Name(), "error", err.Error())
+				}
+			}(exporter)
+		}
+
+		wg.Wait()
+		w.WriteHeader(http.StatusAccepted)
+		_, _ = w.Write([]byte("ok"))
+	})
+
+	if len(ar.config.CORSAllowedOrigins) > 0 {
+		c := cors.New(cors.Options{
+			AllowedOrigins: ar.config.CORSAllowedOrigins,
+		})
+		handler = c.Handler(handler)
+	}
+
+	return handler
+}