+++ pkg/integrations/v2/app_o11y_receiver/models/payload.go
@@ -1,0 +1,9 @@
+package models
+
+// Payload is the body of the receiver request
+type Payload struct {
+	Exceptions   []Exception   `json:"exceptions,omitempty"`
+	Logs         []Log         `json:"logs,omitempty"`
+	Measurements []Measurement `json:"measurements,omitempty"`
+	Meta         Meta          `json:"meta,omitempty"`
+}