+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This release has deprecations. Please read [DEPRECATION] entries and consult
 the [upgrade guide](https://github.com/grafana/agent/blob/main/docs/upgrade-guide/_index.md)
 for detailed information.
 
+- [FEATURE] (beta) Add app o11y integration. This depends on integrations-next being enabled
+   via the `integrations-next` feature flag. Use `-enable-features=integrations-next` to used
+   this integration. (@kpelelis)
+
 - [FEATURE] (beta) Enable experimental config urls for fetching remote configs. Currently,
    only HTTP/S is supported. Pass the `-enable-features=remote-configs` flag to turn this on. (@rlankfo)